refactor file names, location, and contents
