+++ test/all.dart
@@ -3,20 +3,22 @@
 
 library all_test;
 
+import 'grinder_test.dart' as grinder_test;
+import 'grinder_files_test.dart' as grinder_files_test;
+import 'grinder_sdk_test.dart' as grinder_sdk_test;
+import 'grinder_tools_test.dart' as grinder_tools_test;
+import 'integration_test.dart' as integration_test;
 import 'src/cli_test.dart' as src_cli_test;
 import 'src/discover_tasks_test.dart' as src_discover_tasks_test;
 import 'src/utils_test.dart' as src_utils_test;
-import 'cli_test.dart' as cli_test;
-import 'grinder_test.dart' as grinder_test;
-import 'grinder_files_test.dart' as grinder_files_test;
-import 'grinder_tools_test.dart' as grinder_tools_test;
 
 main() {
+  grinder_test.main();
+  grinder_files_test.main();
+  grinder_sdk_test.main();
+  grinder_tools_test.main();
+  integration_test.main();
   src_cli_test.main();
   src_discover_tasks_test.main();
   src_utils_test.main();
-  cli_test.main();
-  grinder_test.main();
-  grinder_files_test.main();
-  grinder_tools_test.main();
 }
