+++ test/src/_common.dart
@@ -1,7 +1,7 @@
 // Copyright 2014 Google. All rights reserved. Use of this source code is
 // governed by a BSD-style license that can be found in the LICENSE file.
 
-library grinder_test_utils;
+library grinder.test.common;
 
 import 'dart:async';
