+++ test/src/task_discovery/bad_tasks.dart
