+++ test/src/task_discovery/external_tasks.dart
